Prettier usage indicator
